create separate module for primitives
