+++ interp/src/primitives/mod.rs
@@ -1,0 +1,4 @@
+mod primitive;
+pub use primitive::Primitive;
+
+pub mod combinational;
